instabot_sqs_lambda/requirements.txt
instagrapi==2.1.3
pillow==11.1.0
openpyxl
pandas